DTFS2-3724 - reusable `assign to` select mappings for tasks and lead underwriter assignment
