+++ trade-finance-manager-ui/server/controllers/case/helpers.js
@@ -1,5 +1,3 @@
-import { userFullName } from '../../helpers/user';
-
 export const getGroup = (groupId, allTasks) =>
   allTasks.find((group) => group.id === groupId);
 
@@ -17,104 +15,4 @@ export const getTask = (groupId, taskId, tasks) => {
   }
 
   return null;
-};
-
-export const isTaskAssignedToUser = (assignedToUserId, userId) => {
-  if (assignedToUserId === userId) {
-    return true;
-  }
-
-  return false;
-};
-
-export const getTeamMembersWithoutCurrentUser = (teamMembers, currentUserId) =>
-  teamMembers.filter((teamMember) =>
-    teamMember._id !== currentUserId); // eslint-disable-line no-underscore-dangle
-
-
-export const mapTeamMembersSelectOptions = (members, taskAssignedTo, currentUserId) => {
-  const membersWithoutCurrentUser = getTeamMembersWithoutCurrentUser(members, currentUserId);
-
-  return membersWithoutCurrentUser.map((member) => {
-    // eslint-disable-next-line no-underscore-dangle
-    const { _id: memberId } = member;
-
-    return {
-      value: memberId,
-      text: userFullName(member),
-      selected: taskAssignedTo === memberId,
-    };
-  });
-};
-
-export const mapAssignToSelectOptions = (task, currentUser, allTeamMembers) => {
-  const taskAssignedTo = task.assignedTo.userId;
-  // eslint-disable-next-line no-underscore-dangle
-  const taskIsAssignedToUser = isTaskAssignedToUser(taskAssignedTo, currentUser._id);
-
-  const currentUserFullName = userFullName(currentUser);
-  const assignToMeCopy = `${currentUserFullName} (Assign to me)`;
-
-  const currentUserId = currentUser._id; // eslint-disable-line no-underscore-dangle
-
-  const mappedTeamMembersSelectOptions = mapTeamMembersSelectOptions(
-    allTeamMembers,
-    taskAssignedTo,
-    currentUserId,
-  );
-
-  const taskIsUnassigned = task.assignedTo.userId === 'Unassigned';
-
-  // 3 possible states:
-  // task is assigned to someone that is not the currently logged in user.
-  // task is unassigned
-  // task is assigned to me/currently logged in user.
-
-  let mapped = [
-    {
-      value: currentUserId,
-      text: assignToMeCopy,
-      selected: false,
-    },
-    {
-      value: 'Unassigned',
-      text: 'Unassigned',
-      selected: false,
-    },
-    ...mappedTeamMembersSelectOptions,
-  ];
-
-  if (taskIsUnassigned) {
-    mapped = [
-      {
-        value: 'Unassigned',
-        text: 'Unassigned',
-        selected: true,
-      },
-      {
-        value: currentUserId,
-        text: assignToMeCopy,
-        selected: false,
-      },
-      ...mappedTeamMembersSelectOptions,
-    ];
-  }
-
-  if (taskIsAssignedToUser) {
-    mapped = [
-      {
-        value: 'Unassigned',
-        text: 'Unassigned',
-        selected: false,
-      },
-      {
-        value: currentUserId,
-        text: currentUserFullName,
-        selected: true,
-      },
-      ...mappedTeamMembersSelectOptions,
-    ];
-  }
-
-  return mapped;
 };
