+++ trade-finance-manager-ui/server/controllers/case/underwriting/index.js
@@ -1,8 +1,10 @@
+import leadUnderwriter from './lead-underwriter';
 import pricingAndRisk from './pricing-and-risk';
 import bankSecurity from './bank-security';
 import underwriterManagersDecision from './underwriter-managers-decision';
 
 export default {
+  ...leadUnderwriter,
   ...pricingAndRisk,
   ...bankSecurity,
   ...underwriterManagersDecision,
