+++ trade-finance-manager-ui/server/helpers/user.js
@@ -5,3 +5,12 @@ export const userFullName = (user) => {
 
 export const userIsInTeam = (user, teamList) =>
   teamList.some((teamId) => user.teams.includes(teamId));
+
+
+export const isAssignedToUser = (assignedToUserId, userId) => {
+  if (assignedToUserId === userId) {
+    return true;
+  }
+
+  return false;
+};
